test(harness): share Agent Map projection fixtures
Closes: SAP-3064
